+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Added a guard that allows to deselect the chip on drop.
+
 ## [0.1.0] - 2024-07-06
 
 [v0.1.0]: https://github.com/sunsided/chip-select/releases/tag/v0.1.0